[refactor] Moved tests related to tcp and udp port fixtures to a separate module
